+++ tests/conftest.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
-from typing import Callable, List, Tuple
+from typing import Callable
 
 from aiohttp.test_utils import TestClient
 from aiohttp.web import Application
@@ -9,14 +8,10 @@ from pytest import fixture
 from sqlitemap import Connection
 
 from iftttie import templates
+from iftttie.automation import Automation
 from iftttie.web import Context, routes
 
 pytest_plugins = 'aiohttp.pytest_plugin'
-
-
-@dataclass
-class Setup:
-    USERS: List[Tuple[str, str]]
 
 
 @fixture
@@ -26,9 +21,10 @@ async def db() -> Connection:
 
 @fixture
 async def client(aiohttp_client: Callable[..., TestClient], db: Connection) -> TestClient:
-    users = [('test', '$pbkdf2-sha256$29000$tPaeE.Kck5Ly/n/vnXOuFQ$2C/q3Fk/5MoX149flqyXow0CY/UCSt31Ga4xIZAWeEg')]
+    automation = Automation()
+    automation.users.append(('test', '$pbkdf2-sha256$29000$tPaeE.Kck5Ly/n/vnXOuFQ$2C/q3Fk/5MoX149flqyXow0CY/UCSt31Ga4xIZAWeEg'))
     app = Application()
-    app['context'] = Context(setup=Setup(USERS=users), db=db)
+    app['context'] = Context(automation=automation, db=db)
     app.add_routes(routes)
     templates.setup(app)
     return await aiohttp_client(app)